# Downloads/tp5-DAI/tp5-DAI/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mini Pokédex API",
    "description": "Servidor proxy que expone endpoints de PokéAPI para la aplicación Mini Pokédex Web. Permite consultar datos de Pokémon, tipos, habilidades, especies, encuentros y generaciones.",
    "version": "1.0.0",
    "contact": {
      "name": "TP5 DAI"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/pokemon": {
      "get": {
        "tags": [
          "Pokemon"
        ],
        "summary": "Obtiene una lista paginada de Pokémon",
        "description": "Devuelve un listado de Pokémon con su nombre y URL de detalle. La cantidad y posición se controlan con los parámetros limit y offset.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Cantidad de Pokémon a devolver. Por defecto: 20. Máximo recomendado: 100.",
            "required": false,
            "type": "integer",
            "example": 20
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Posición desde la que empezar la lista (para paginación). Por defecto: 0.",
            "required": false,
            "type": "integer",
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de Pokémon obtenida correctamente.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "example": {
                          "type": "number",
                          "example": 1302
                        },
                        "description": {
                          "type": "string",
                          "example": "Total de Pokémon disponibles en la API"
                        }
                      }
                    },
                    "next": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "https://pokeapi.co/api/v2/pokemon?offset=20&limit=20"
                        },
                        "description": {
                          "type": "string",
                          "example": "URL de la siguiente página (null si no hay más)"
                        }
                      }
                    },
                    "previous": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {},
                        "description": {
                          "type": "string",
                          "example": "URL de la página anterior (null si es la primera)"
                        }
                      }
                    },
                    "results": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "array"
                        },
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string",
                              "example": "object"
                            },
                            "properties": {
                              "type": "object",
                              "properties": {
                                "name": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "bulbasaur"
                                    }
                                  }
                                },
                                "url": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "string"
                                    },
                                    "example": {
                                      "type": "string",
                                      "example": "https://pokeapi.co/api/v2/pokemon/1/"
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor.",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "Error interno del servidor"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}